+++ backend/garpix_user/tasks/__init__.py
@@ -1,1 +1,2 @@
 from .delete_unconfirmed_users import delete_unconfirmed_users  # noqa
+from .password_valifity_passed import password_valifity_passed  # noqa